refactoring utils and abstractions
