+++ dynamax/nonlinear_gaussian_ssm/nonlinear_gaussian_ssm.py
@@ -3,7 +3,7 @@ from typing import NamedTuple, Optional, Union, Callable
 import tensorflow_probability.substrates.jax as tfp
 from tensorflow_probability.substrates.jax.distributions import MultivariateNormalFullCovariance as MVN
 import tensorflow_probability.substrates.jax.distributions as tfd
-from dynamax.abstractions import SSM
+from dynamax.ssm import SSM
 from dynamax.linear_gaussian_ssm.inference import PosteriorLGSSMFiltered, PosteriorLGSSMSmoothed
 
 tfd = tfp.distributions
