+++ dynamax/ssm.py
